# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ApiResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TestResponse": {
				"properties": {
					"test_id": {
						"type": "string"
					},
					"course_id": {
						"type": "string"
					},
					"type_status": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"test_id",
					"type_status",
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TestResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/TestResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TestResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/TestResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTestRequest": {
				"properties": {
					"course_id": {
						"type": "string"
					},
					"type_status": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"type_status",
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateTestRequest": {
				"properties": {
					"test_id": {
						"type": "string"
					},
					"course_id": {
						"type": "string"
					},
					"type_status": {
						"type": "string"
					},
					"title": {
						"type": "string"
					}
				},
				"required": [
					"test_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SectionResponse": {
				"properties": {
					"section_id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"chapter_id": {
						"type": "string"
					},
					"video_url": {
						"type": "string"
					},
					"knowledge_points": {
						"type": "string"
					},
					"video_subtitles": {
						"type": "string"
					},
					"knowledge_content": {
						"type": "string"
					},
					"estimated_time": {
						"type": "number",
						"format": "double"
					},
					"section_order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"section_id",
					"title",
					"chapter_id",
					"section_order"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_SectionResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/SectionResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_SectionResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/SectionResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateSectionRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"chapter_id": {
						"type": "string"
					},
					"video_url": {
						"type": "string"
					},
					"knowledge_points": {
						"type": "string"
					},
					"video_subtitles": {
						"type": "string"
					},
					"knowledge_content": {
						"type": "string"
					},
					"estimated_time": {
						"type": "number",
						"format": "double"
					},
					"section_order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"title",
					"chapter_id",
					"section_order"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateSectionRequest": {
				"properties": {
					"section_id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"chapter_id": {
						"type": "string"
					},
					"video_url": {
						"type": "string"
					},
					"knowledge_points": {
						"type": "string"
					},
					"video_subtitles": {
						"type": "string"
					},
					"knowledge_content": {
						"type": "string"
					},
					"estimated_time": {
						"type": "number",
						"format": "double"
					},
					"section_order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"section_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LearningRecordResponse": {
				"properties": {
					"task_id": {
						"type": "string"
					},
					"plan_id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"section_id": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"task_id",
					"plan_id",
					"user_id",
					"section_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_LearningRecordResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/LearningRecordResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_LearningRecordResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"$ref": "#/components/schemas/LearningRecordResponse"
					},
					"message": {
						"type": "string"
					},
					"error": {
						"type": "string"
					},
					"details": {},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateLearningRecordRequest": {
				"properties": {
					"plan_id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"section_id": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"plan_id",
					"user_id",
					"section_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateLearningRecordRequest": {
				"properties": {
					"task_id": {
						"type": "string"
					},
					"plan_id": {
						"type": "string"
					},
					"user_id": {
						"type": "string"
					},
					"section_id": {
						"type": "string"
					},
					"start_date": {
						"type": "string",
						"format": "date-time"
					},
					"end_date": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string"
					}
				},
				"required": [
					"task_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "express-pg-orm-backend-ts",
		"version": "1.0.0",
		"description": "TypeScript Express.js backend with PostgreSQL and Sequelize ORM",
		"contact": {}
	},
	"paths": {
...
		"/tests/list": {
			"post": {
				"operationId": "ListTests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TestResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/tests/getById": {
			"post": {
				"operationId": "GetTestById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TestResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"test_id": {
										"type": "string"
									}
								},
								"required": [
									"test_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/tests/add": {
			"post": {
				"operationId": "AddTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTestRequest"
							}
						}
					}
				}
			}
		},
		"/tests/update": {
			"post": {
				"operationId": "UpdateTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTestRequest"
							}
						}
					}
				}
			}
		},
		"/tests/delete": {
			"post": {
				"operationId": "DeleteTest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"测试表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"test_id": {
										"type": "string"
									}
								},
								"required": [
									"test_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/sections/list": {
			"post": {
				"operationId": "ListSections",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_SectionResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"节表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/sections/getSectionById": {
			"post": {
				"operationId": "GetSectionById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_SectionResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"节表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"section_id": {
										"type": "string"
									}
								},
								"required": [
									"section_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/sections/add": {
			"post": {
				"operationId": "AddSection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"节表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSectionRequest"
							}
						}
					}
				}
			}
		},
		"/sections/update": {
			"post": {
				"operationId": "UpdateSection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"节表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSectionRequest"
							}
						}
					}
				}
			}
		},
		"/sections/delete": {
			"post": {
				"operationId": "DeleteSection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"节表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"section_id": {
										"type": "string"
									}
								},
								"required": [
									"section_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/learning-records/list": {
			"post": {
				"operationId": "ListLearningRecords",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_LearningRecordResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"学习记录表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"limit": {
										"type": "number",
										"format": "double"
									},
									"page": {
										"type": "number",
										"format": "double"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/learning-records/getLearningRecordById": {
			"post": {
				"operationId": "GetLearningRecordById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_LearningRecordResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"学习记录表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"task_id": {
										"type": "string"
									}
								},
								"required": [
									"task_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/learning-records/add": {
			"post": {
				"operationId": "AddLearningRecord",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"学习记录表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLearningRecordRequest"
							}
						}
					}
				}
			}
		},
		"/learning-records/update": {
			"post": {
				"operationId": "UpdateLearningRecord",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"学习记录表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLearningRecordRequest"
							}
						}
					}
				}
			}
		},
		"/learning-records/delete": {
			"post": {
				"operationId": "DeleteLearningRecord",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"学习记录表"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"task_id": {
										"type": "string"
									}
								},
								"required": [
									"task_id"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}